OM-50871: RB:34483 Changes in CBTP cost dtos to support zonal RIs.

+++ com.vmturbo.analysis/platform.analysis/src/main/java/com/vmturbo/platform/analysis/utilities/Quote.java
@@ -341,7 +341,7 @@ public abstract class Quote {
 
         protected CommodityCloudQuote(@Nullable final Trader seller,
                                       final double quoteValue,
-                                      @Nullable final Integer regionId) {
+                                      @Nullable final Long regionId) {
             super(seller, quoteValue);
             MoveContext.Builder builder = MoveContext.newBuilder();
             if (regionId != null) {
